refactor: Update ReactDOM import and use createRoot for rendering
This commit updates the import statement for ReactDOM in the index.js file to use the new createRoot function from the react-dom/client module. It also wraps the App component with the Provider component from react-redux to provide the Redux store to the entire application.

The changes improve the performance of the application by using the createRoot function for rendering and enabling the use of Redux for state management.