refactor: Fix bug in cartSlice.js addItem function
This commit fixes a bug in the addItem function of cartSlice.js. The bug was causing the existingItem's totalPrice to be incorrectly calculated when adding a new item to the cart. The bug was fixed by using the newItem's price instead of the existingItem's price.

The changes ensure that the totalPrice is accurately calculated when adding items to the cart.