refactor: Update Cart, CartButton, CartItem, ProductItem, and Products components
This commit updates multiple components in the codebase. The Cart component now uses the useSelector hook from react-redux to retrieve items from the cart state. The CartItem component has been modified to dispatch actions for removing and adding items to the cart. The CartButton component now displays the total quantity of items in the cart by using the useSelector hook. The ProductItem component dispatches an action to add items to the cart. Lastly, the Products component now dynamically renders a list of products using the DUMMY_PRODUCTS data.

These changes improve the functionality and user experience of the shopping cart feature.