refactor: Update App.js and ProductItem.js to send cart data to Firebase
This commit modifies the App.js and ProductItem.js files to send the cart data to Firebase when the cart is updated. The useEffect hook is used in App.js to make a PUT request to the Firebase database with the updated cart data. In ProductItem.js, the addToCartHandler function is updated to include a comment indicating that an HTTP request to the Firebase URL should be made to update the cart.

These changes improve the functionality of the application by ensuring that the cart data is persisted in the Firebase database.