refactor: Separate cart data sending logic into cartActions
This commit refactors the code to separate the responsibility of sending cart data from the App component. The cart data sending logic is moved to the cartActions file, where two new functions, fetchCartData and sendCartData, are defined. The fetchCartData function makes a GET request to retrieve the cart data from the Firebase database, while the sendCartData function makes a PUT request to update the cart data in the database. This change improves code organization and maintainability by encapsulating the cart data sending logic in a separate module.